perf improvements

+++ mog/size.mojo
@@ -4,7 +4,7 @@ from mog._extensions import NEWLINE
 from mog._properties import Dimensions
 
 
-def get_width[origin: ImmOrigin, //](text: StringSpan[origin]) -> UInt:
+def get_width[origin: ImmOrigin, //](text: StringSpan[origin]) -> UInt16:
     """Returns the cell width of characters in the string. ANSI sequences are
     ignored and characters wider than one cell (such as Chinese characters and
     emojis) are appropriately measured.
@@ -17,16 +17,16 @@ def get_width[origin: ImmOrigin, //](text: StringSpan[origin]) -> UInt:
     Returns:
         The width of the string in cells.
     """
-    var width: UInt = 0
+    var width: UInt16 = 0
     for line in text.splitlines():
-        var w = ansi.printable_rune_width(line)
+        var w = UInt16(ansi.printable_rune_width(line))
         if w > width:
             width = w
 
     return width
 
 
-def get_height[origin: ImmOrigin, //](text: StringSpan[origin]) -> UInt:
+def get_height[origin: ImmOrigin, //](text: StringSpan[origin]) -> UInt16:
     """Returns height of a string in cells. This is done simply by
     counting \\n characters. If your strings use \\r\\n for newlines you should
     convert them to \\n first, or simply write a separate function for measuring
@@ -38,7 +38,7 @@ def get_height[origin: ImmOrigin, //](text: StringSpan[origin]) -> UInt:
     Returns:
         The height of the string in cells.
     """
-    return UInt(text.count(NEWLINE) + 1)
+    return UInt16(text.count(NEWLINE) + 1)
 
 
 def get_dimensions[origin: ImmOrigin, //](text: StringSpan[origin]) -> Dimensions:
@@ -52,4 +52,4 @@ def get_dimensions[origin: ImmOrigin, //](text: StringSpan[origin]) -> Dimension
     Returns:
         The width and height of the string in cells.
     """
-    return Dimensions(width=UInt16(get_width(text)), height=UInt16(get_height(text)))
+    return Dimensions(width=get_width(text), height=get_height(text))
